more readable on small displays
this isn't perfect -- poems have a nowrap. this is to prevent them from wrapping
and being confused with the subsequent line. so many of the poems have indented
lines (distinct from wrapping)

there are a couple of poems in the book with both dropped lines and wrapping,
leaving something looking like:

  The first line has some
      words,
    Second with more.
  The third's for the birds,
    Fourth a score.

i coded each stanza with <p>, the indents with &emsp; and forced line
breaks with <br/>. this makes line wrapping difficult.